Refactor store rest controller to optimize database access
fixes #490
