+++ spring-content-rest/src/main/java/internal/org/springframework/content/rest/contentservice/ContentServiceFactory.java
@@ -1,18 +1,13 @@
-package internal.org.springframework.content.rest.controllers;
+package internal.org.springframework.content.rest.contentservice;
 
 import org.springframework.content.commons.repository.AssociativeStore;
 import org.springframework.content.commons.repository.ContentStore;
-import org.springframework.content.commons.repository.Store;
-import org.springframework.content.commons.storeservice.StoreInfo;
 import org.springframework.content.commons.storeservice.Stores;
 import org.springframework.content.rest.config.RestConfiguration;
-import org.springframework.content.rest.controllers.ContentService;
 import org.springframework.data.repository.support.Repositories;
 import org.springframework.data.repository.support.RepositoryInvokerFactory;
 
-import internal.org.springframework.content.rest.controllers.ContentServiceHandlerMethodArgumentResolver.ContentStoreContentService;
-import internal.org.springframework.content.rest.controllers.ContentServiceHandlerMethodArgumentResolver.StoreContentService;
-import internal.org.springframework.content.rest.io.AssociatedResource;
+import internal.org.springframework.content.rest.io.AssociatedStoreResource;
 import internal.org.springframework.content.rest.io.StoreResource;
 import internal.org.springframework.content.rest.mappings.StoreByteRangeHttpRequestHandler;
 
@@ -32,18 +27,19 @@ public class ContentServiceFactory {
         this.byteRangeRestRequestHandler = byteRangeRestRequestHandler;
     }
 
-    public ContentService getContentService(StoreInfo storeInfo, StoreResource resource) {
+    public ContentService getContentService(StoreResource resource) {
 
-        if (ContentStore.class.isAssignableFrom(storeInfo.getInterface())) {
+        if (ContentStore.class.isAssignableFrom(resource.getStoreInfo().getInterface())) {
 
-            Object entity = ((AssociatedResource)resource).getAssociation();
-            return new ContentStoreContentService(config, storeInfo, repoInvokerFactory.getInvokerFor(entity.getClass()), entity, byteRangeRestRequestHandler);
-        } else if (AssociativeStore.class.isAssignableFrom(storeInfo.getInterface())) {
+            Object entity = ((AssociatedStoreResource)resource).getAssociation();
+
+            return new ContentStoreContentService(config, null, repoInvokerFactory.getInvokerFor(entity.getClass()), entity, byteRangeRestRequestHandler);
+        } else if (AssociativeStore.class.isAssignableFrom(resource.getStoreInfo().getInterface())) {
 
             throw new UnsupportedOperationException("AssociativeStore not supported");
         } else {
 
-            return new StoreContentService(storeInfo.getImplementation(Store.class), byteRangeRestRequestHandler);
+            return new StoreContentService(byteRangeRestRequestHandler);
         }
     }
 }
