+++ spring-content-rest/src/main/java/internal/org/springframework/content/rest/controllers/ResourceHandlerMethodArgumentResolver.java
@@ -2,7 +2,6 @@ package internal.org.springframework.content.rest.controllers;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.springframework.content.commons.renditions.Renderable;
 import org.springframework.content.commons.repository.AssociativeStore;
 import org.springframework.content.commons.repository.Store;
 import org.springframework.content.commons.storeservice.StoreInfo;
@@ -17,9 +16,8 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import org.springframework.web.util.UrlPathHelper;
 
-import internal.org.springframework.content.rest.io.AssociatedResource;
-import internal.org.springframework.content.rest.io.AssociatedResourceImpl;
-import internal.org.springframework.content.rest.io.RenderableResourceImpl;
+import internal.org.springframework.content.rest.io.AssociatedStorePropertyResourceImpl;
+import internal.org.springframework.content.rest.io.AssociatedStoreResourceImpl;
 import internal.org.springframework.content.rest.io.StoreResourceImpl;
 import internal.org.springframework.content.rest.utils.StoreUtils;
 
@@ -45,18 +43,14 @@ public class ResourceHandlerMethodArgumentResolver extends StoreHandlerMethodArg
       String path = new UrlPathHelper().getPathWithinApplication(nativeWebRequest.getNativeRequest(HttpServletRequest.class));
       String pathToUse = path.substring(StoreUtils.storePath(info).length() + 1);
 
-      return new StoreResourceImpl(info.getImplementation(Store.class).getResource(pathToUse));
+      return new StoreResourceImpl(info, info.getImplementation(Store.class).getResource(pathToUse));
     }
 
     @Override
     protected Object resolveAssociativeStoreEntityArgument(StoreInfo storeInfo, Object domainObj) {
 
         Resource r = storeInfo.getImplementation(AssociativeStore.class).getResource(domainObj);
-        r = new AssociatedResourceImpl(domainObj, r);
-
-        if (Renderable.class.isAssignableFrom(storeInfo.getInterface())) {
-            r = new RenderableResourceImpl((Renderable)storeInfo.getImplementation(AssociativeStore.class), (AssociatedResource)r);
-        }
+        r = new AssociatedStoreResourceImpl(storeInfo, domainObj, r);
         return r;
     }
 
@@ -65,11 +59,7 @@ public class ResourceHandlerMethodArgumentResolver extends StoreHandlerMethodArg
 
         AssociativeStore s = storeInfo.getImplementation(AssociativeStore.class);
         Resource resource = s.getResource(propertyVal);
-//        resource = new AssociatedResourceImpl(propertyVal, resource);
-        resource = new AssociatedResourceImpl(propertyVal, domainObj, resource);
-        if (Renderable.class.isAssignableFrom(storeInfo.getInterface())) {
-            resource = new RenderableResourceImpl((Renderable)storeInfo.getImplementation(AssociativeStore.class), (AssociatedResource)resource);
-        }
+        resource = new AssociatedStorePropertyResourceImpl(storeInfo, propertyVal, embeddedProperty, domainObj, resource);
         return resource;
     }
 }
