+++ spring-content-rest/src/main/java/internal/org/springframework/content/rest/io/AssociatedStorePropertyResource.java
@@ -1,0 +1,8 @@
+package internal.org.springframework.content.rest.io;
+
+public interface AssociatedStorePropertyResource<S> extends AssociatedStoreResource<S> {
+
+    public boolean embedded();
+
+    public Object getProperty();
+}
