+++ spring-content-rest/src/main/java/internal/org/springframework/content/rest/io/AssociatedStoreResource.java
@@ -2,7 +2,8 @@ package internal.org.springframework.content.rest.io;
 
 import org.springframework.core.io.WritableResource;
 
-public interface AssociatedResource<S> extends WritableResource, StoreResource {
+public interface AssociatedStoreResource<S> extends WritableResource, StoreResource {
 
     S getAssociation();
+
 }
